libdot: storage: add helper for generating change events
Most of our storage backends all have their own ad-hoc implementation
for generating change events to try and mimic what Chrome itself does.
Lets add a single implementation to avoid the duplication, and that is
better at matching the Chrome behavior.

The biggest thing is omitting the oldValue/newValue properties when
they're undefined which is what Chrome does -- our code initializes
both all the time to an explicit undefined which confuses code that
does property tests rather than dereferencing it.

There is some downside in that we have to fully construct both of
the storage states (before & after) in order to generate the events
where some of the code currently can avoid that overhead, but these
APIs should not be performance critical or storing a ton of data,
so the overhead shouldn't be that significant in the real world.

Change-Id: Idf24df5cb7d80c7615e4f63f0e4febd8c5d9932e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2269586
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>