kokoro: update container with more packages
New Chrome releases require more DRM packages to run.

While testing the new build, noticed the kokoro build logic broke
due to the shell=True usage.  Rework the code to use Python so we
avoid both shell commands & the breakage in general.

The newer version of pylint also seems to have fixed the recursion
errors we were seeing with some modules, and it triggers a warning
in existing libdot code.  So clean that up too.

Change-Id: I4ebbc87a6a5a839cb221727bb71877fac1e63f5f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2473678
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>